+++ releasenotes.md
@@ -29,6 +29,8 @@ Release Notes
   used. For example, `gmali.com` will become `gmail.com`.
 * When `hashAddress` is used, all trailing periods are now removed from an
   email address domain. Previously only a single period was removed.
+* When `hashAddress` is used, the local part of an email address is now
+  normalized to NFC.
 
 4.2.0 (2023-12-05)
 ------------------